Add a first stab at a basic-apue-skills-assessment script.
The class CS631 Advanced Programming in the Unix Environment
(http://www.cs.stevens.edu/~jschauma/631/) has no strict pre-requisites,
but I need to assess whether a prospective student might be completely
lost and not likely to be able to perform even the most basic of tasks
required to follow the course.  I use this script to see if students are
able to operate within a unix-like environment and know how to at least
compile a simple program.

The script prompts the user to perform certain tasks; all interactions
are logged via script(1), and the output is stored in the instructor's
(world-writable) directory.  (We don't email the output, since outgoing
mail at this institution goes through Google, and the outgoing mail
filter rejects executables.  We can't reliably remove any executables
students might (accidentally or intentionally) create.)