email output instead of copying it on the local file systems
We originally tried to email a tar of all files, but that might have
included executables created by the students, and the outgoing mail
server might not allow you to send executables (*cough* gmail/google
apps *cough*).  But it just so happens that the instructor's directory
is on a different network share from the directories of the students,
and students may not be able to automount it.  So we now go back to
emailing the output, but only email the actual script files.